requirements.txt
python-dotenv
streamlit
together
pandas